Move model_loader.py to api directory and update import
Renamed model_loader.py from src/ to src/api/ and updated the import path in main.py
